requirements.txt
# 2019-07-27
#adafruit_servokit
Flask
Flask-RESTful
PyYAML
jsons
